pyproject.toml
[tool.poetry]
name = "aldotaibot"
version = "2.3"
description = ""
authors = ["Your Name <you@example.com>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.3"
creart = "^0.2.2"
html2text = "^2020.1.16"
loguru = "^0.6.0"
psutil = "^5.9.4"
pydoodle = "^1.1.3"
pyjson5 = "^1.6.2"
graia-ariadne = "^0.10.0"
graia-saya = "^0.0.17"
graia-scheduler = "^0.0.10"
arclet-alconna = "^1.4.2.1"
pillow = "^9.3.0"
nest-asyncio = "^1.5.6"
filetype = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
